pyproject.toml
[project]
name = "wallets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "django-filter>=25.1",
    "djangorestframework>=3.15.2",
    "djangorestframework-jsonapi>=7.1.0",
    "dotenv>=0.9.9",
    "environs>=14.1.1",
    "psycopg-pool>=3.2.6",
    "psycopg[binary]>=3.2.6",
]

[dependency-groups]
dev = [
    "coverage>=7.7.1",
    "django-stubs[compatible-mypy]>=5.1.3",
    "djangorestframework-stubs[compatible-mypy]>=3.15.3",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "pytest-factoryboy>=2.7.0",
    "ruff>=0.11.2",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "TID", # flake8-tidy-imports
    "T20", # flake8-print
    "I",   # isort
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings"

[[tool.mypy.overrides]]
module = "rest_framework_json_api.*"
ignore_missing_imports = true
